+++ src/auth.h
@@ -41,7 +41,7 @@ void    auth_display(AUTH this, SCHEME scheme);
 char *  auth_get_basic_header(AUTH this, SCHEME scheme);
 BOOLEAN auth_set_basic_header(AUTH this, SCHEME scheme, char *realm);
 char *  auth_get_digest_header(AUTH this, SCHEME scheme, DCHLG *chlg, DCRED *cred, const char *meth, const char *uri);
-BOOLEAN auth_set_digest_header(AUTH this, DCHLG **ch, DCRED **cr, size_t *rand, char *realm, char *str);
+BOOLEAN auth_set_digest_header(AUTH this, DCHLG **ch, DCRED **cr, unsigned int *rand, char *realm, char *str);
 BOOLEAN auth_get_proxy_required(AUTH this);
 void    auth_set_proxy_required(AUTH this, BOOLEAN required);
 char *  auth_get_proxy_host(AUTH this);
